Consolidate Sanity live API
